+++ elyra/util/github.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2018-2021 Elyra Authors
+# Copyright 2018-2022 Elyra Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -88,9 +88,9 @@ class GithubClient(LoggingConfigurable):
                                'Please validate your runtime configuration details and retry.') from e
 
     @staticmethod
-    def get_github_url(api_url: str,
-                       repository_name: str,
-                       repository_branch: str) -> str:
+    def get_git_url(api_url: str,
+                    repository_name: str,
+                    repository_branch: str) -> str:
         """
         Generates the URL to the location of the pushed DAG
         :param api_url: url of the GitHub API