+++ packages/r-editor/src/REditor.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2021 Elyra Authors
+ * Copyright 2018-2022 Elyra Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,14 +15,8 @@
  */
 import { ScriptEditor } from '@elyra/script-editor';
 
-import { CodeEditor, IEditorServices } from '@jupyterlab/codeeditor';
-import {
-  ABCWidgetFactory,
-  DocumentRegistry,
-  DocumentWidget
-} from '@jupyterlab/docregistry';
+import { DocumentRegistry, DocumentWidget } from '@jupyterlab/docregistry';
 import { FileEditor } from '@jupyterlab/fileeditor';
-
 import { Widget } from '@lumino/widgets';
 
 export class REditor extends ScriptEditor {
@@ -33,67 +27,13 @@ export class REditor extends ScriptEditor {
     options: DocumentWidget.IOptions<FileEditor, DocumentRegistry.ICodeModel>,
     getCurrentWidget: () => Widget | null
   ) {
-    super(options, 'R', getCurrentWidget);
-
-    // Add icon to main tab
-    this.title.icon = 'rIcon';
+    super(options, getCurrentWidget);
   }
-}
-
-/**
- * A widget factory for R Editors.
- */
-export class REditorFactory extends ABCWidgetFactory<
-  REditor,
-  DocumentRegistry.ICodeModel
-> {
-  /**
-   * Construct a new editor widget factory.
-   */
-  constructor(options: REditorFactory.IOptions) {
-    super(options.factoryOptions);
-    this._services = options.editorServices;
-    this.getCurrentWidget = options.getCurrentWidget;
+  getLanguage(): string {
+    return 'R';
   }
 
-  /**
-   * Create a new widget given a context.
-   */
-  protected createNewWidget(context: DocumentRegistry.CodeContext): REditor {
-    const newDocumentEditor = this._services.factoryService.newDocumentEditor;
-    const factory: CodeEditor.Factory = options => {
-      return newDocumentEditor(options);
-    };
-    const content = new FileEditor({
-      factory,
-      context,
-      mimeTypeService: this._services.mimeTypeService
-    });
-    return new REditor({ content, context }, this.getCurrentWidget);
-  }
-
-  private _services: IEditorServices;
-  private getCurrentWidget: () => Widget | null;
-}
-
-/**
- * The namespace for `REditorFactory` class statics.
- */
-export namespace REditorFactory {
-  /**
-   * The options used to create an editor widget factory.
-   */
-  export interface IOptions {
-    /**
-     * The editor services used by the factory.
-     */
-    editorServices: IEditorServices;
-
-    /**
-     * The factory options associated with the factory.
-     */
-    factoryOptions: DocumentRegistry.IWidgetFactoryOptions<REditor>;
-
-    getCurrentWidget: () => Widget | null;
+  getIcon(): string {
+    return 'rIcon';
   }
 }